requirements.txt
absl-py==2.0.0
alabaster==0.7.13
anyascii==0.3.2
astroid==3.0.2
astunparse==1.6.3
Babel==2.13.1
build==1.0.3
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
docutils==0.20.1
Flask==3.0.1
Flask-SQLAlchemy==3.1.1
flatbuffers==23.5.26
gast==0.5.4
google-auth==2.24.0
google-auth-oauthlib==1.1.0
google-pasta==0.2.0
grpcio==1.59.3
h5py==3.10.0
idna==3.6
imagesize==1.4.1
iniconfig==2.0.0
Jinja2==3.1.2
kafka-python==2.0.2
keras==2.15.0
libclang==16.0.6
linkify-it-py==2.0.2
mailparse==1.0.13
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdit-py-plugins==0.4.0
mdurl==0.1.2
ml-dtypes==0.2.0
myst-parser==2.0.0
mysqlclient==2.2.1
mysql-connector-python==8.2.0
numpy==1.26.2
nltk==3.8.1
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.4
pluggy==1.3.0
#protobuf==4.23.4 -- changed protobuf from 4.23.4 to 4.21.12 (version conflicts with mysql-connector-python)
protobuf==4.21.12
pyasn1==0.5.1
pyasn1-modules==0.3.0
Pygments==2.17.2
pyproject_hooks==1.0.0
pytest==7.4.3
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.2
six==1.16.0
snowballstemmer==2.2.0
Sphinx==7.2.6
sphinx-autodoc2==0.5.0
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
typing_extensions==4.8.0
uc-micro-py==1.0.2
urllib3==2.1.0
Werkzeug==3.0.1
wrapt==1.14.1
